+++ osaf/services/infrastructure/dtms/include/dtm_cb.h
@@ -80,7 +80,7 @@ typedef struct dtm_internode_cb {
   DTM_IP_ADDR_TYPE i_addr_family; /* Indicates V4 or V6 */
   uint8_t mcast_flag;     /* Indicates mcast */
   int32_t initial_dis_timeout;
-  int32_t bcast_msg_freq;
+  int64_t bcast_msg_freq;
   NCS_PATRICIA_TREE nodeid_tree;  /* NODE_DB information of Nodes */
   NCS_PATRICIA_TREE comm_sock_tree;       /* NODE_DB information of Nodes */
   NCS_PATRICIA_TREE ip_addr_tree; /* NODE_DB information of Nodes */
